Making vision code more robust to null values

+++ src/main/java/frc/robot/subsystems/Vision/VisionIO.java
@@ -17,7 +17,7 @@ public interface VisionIO {
     public double ty;
     public double tyAdjusted;
     public double tv;
-    public double pipeline;
+    public int pipeline;
     public double tagID;
     public Pose2d estimatedPose;
     public double timestampSeconds;
